fix dbscan test to use parameter
